move preparenode and measurenode to own file
