+++ pennylane/qcut/utils.py
@@ -16,7 +16,6 @@ Support functions for cut_circuit and cut_circuit_mc.
 """
 
 
-import uuid
 import warnings
 from collections.abc import Callable, Sequence
 from typing import Any
@@ -26,47 +25,14 @@ from networkx import MultiDiGraph, has_path, weakly_connected_components
 
 from pennylane import ops
 from pennylane.measurements import MeasurementProcess
-from pennylane.operation import Operation
+from pennylane.ops import Operation
 from pennylane.ops.meta import WireCut
 from pennylane.queuing import WrappedObj
 
 from .cutstrategy import CutStrategy
 from .kahypar import kahypar_cut
+from .ops import MeasureNode, PrepareNode
 from .tapes import graph_to_tape
-
-
-class MeasureNode(Operation):
-    """Placeholder node for measurement operations"""
-
-    num_wires = 1
-    grad_method = None
-    num_params = 0
-
-    def __init__(self, wires=None, id=None):
-        id = id or str(uuid.uuid4())
-
-        super().__init__(wires=wires, id=id)
-
-    def label(self, decimals=None, base_label=None, cache=None):
-        op_label = base_label or self.__class__.__name__
-        return op_label
-
-
-class PrepareNode(Operation):
-    """Placeholder node for state preparations"""
-
-    num_wires = 1
-    grad_method = None
-    num_params = 0
-
-    def __init__(self, wires=None, id=None):
-        id = id or str(uuid.uuid4())
-
-        super().__init__(wires=wires, id=id)
-
-    def label(self, decimals=None, base_label=None, cache=None):
-        op_label = base_label or self.__class__.__name__
-        return op_label
 
 
 def _prep_zero_state(wire):
